Make: architectural changes
