+++ src/pages/Contents/Aside/Aside.scss
