--- src/pages/Contents/Background/Background.scss
@@ -1,26 +1,0 @@
-.background {
-  position: relative;
-  text-align: center;
-  margin: 0 auto;
-
-  &::before {
-    content: '';
-    position: absolute;
-    display: block;
-    width: 100%;
-    height: 400px;
-    top: 0;
-    background: linear-gradient(
-      to right,
-      rgba(34, 34, 34, 1) 0%,
-      rgba(34, 34, 34, 1) 30%,
-      rgba(34, 34, 34, 0.1) 50%,
-      rgba(34, 34, 34, 1) 70%,
-      rgba(34, 34, 34, 1) 100%
-    );
-  }
-
-  .background-image {
-    height: 400px;
-  }
-}
