+++ src/pages/Contents/Contents.scss
@@ -3,12 +3,41 @@
 
   .background {
     position: relative;
+    text-align: center;
+    margin: 0 auto;
     z-index: 1;
+
+    &::before {
+      content: '';
+      position: absolute;
+      display: block;
+      width: 100%;
+      height: 400px;
+      top: 0;
+      background: linear-gradient(
+        to right,
+        rgba(34, 34, 34, 1) 0%,
+        rgba(34, 34, 34, 1) 30%,
+        rgba(34, 34, 34, 0.1) 50%,
+        rgba(34, 34, 34, 1) 70%,
+        rgba(34, 34, 34, 1) 100%
+      );
+    }
+
+    .background-image {
+      height: 400px;
+    }
   }
 
   .main {
     position: relative;
     z-index: 2;
+
+    .main-contents {
+      display: flex;
+      justify-content: center;
+      align-items: flex-start;
+    }
   }
 }
