+++ src/pages/Contents/ContentsThings/BasicInfo.js
@@ -14,7 +14,7 @@ class BasicInfo extends Component {
           </p>
           <p className="little-summary">2021 · 미국 · 액션</p>
           <p className="little-summary">2시간 12분 · 12세</p>
-          <div className="story-summary">
+          <span className="story-summary">
             `텐 링즈`를 차지하는 자, 세상을 지배한다! 초인적인 능력을 가진 텐
             링즈의 힘으로 수세기 동안 어둠의 세상을 지배해 온 `웬우`. `샹치`는
             아버지 웬우 밑에서 암살자로 훈련을 받았지만 이를 거부하고 평범함
@@ -23,7 +23,8 @@ class BasicInfo extends Component {
             내면의 신비한 힘을 일깨우게 된다. 벗어나고 싶은 과거이자, 그
             누구보다 두려운 아버지 웬우를 마주해야 하는 샹치. 악이 될 것인가?
             구원이 될 것인가? 마블의 새로운 시대, 세상에 없던 힘이 탄생한다!
-          </div>
+          </span>
+          <button>더보기</button>
         </article>
       </section>
     );
