+++ src/pages/Contents/ContentsThings/BasicInfo.scss
