--- src/pages/Contents/MainArea/Gallery/Gallery.scss
@@ -1,15 +1,0 @@
-.gallery {
-  margin: 20px 0;
-
-  .title {
-    margin: 20px 20px;
-  }
-
-  &::after {
-    content: '';
-    display: block;
-    margin: 0 auto;
-    width: 300px;
-    border-bottom: 1px solid #f0f0f0;
-  }
-}
