--- src/pages/Contents/MainArea/Gallery/GalleryPictures.scss
@@ -1,23 +1,0 @@
-.gallery-pictures {
-  display: flex;
-  justify-content: flex-start;
-  width: 320px;
-  margin: 10px 0 20px;
-  overflow-x: scroll;
-
-  &::-webkit-scrollbar {
-    display: none;
-  }
-
-  .gallery-picture-sec {
-    margin-left: 10px;
-
-    &:first-child {
-      margin-left: 20px;
-    }
-    .gallery-picture {
-      width: 135px;
-      border-radius: 4px;
-    }
-  }
-}
