--- src/pages/Contents/MainArea/InfoAct/ActStars.scss
@@ -1,11 +1,0 @@
-.star-contents {
-  display: flex;
-  flex-direction: column;
-  margin-left: 20px;
-
-  > span {
-    font-size: 13px;
-    color: #787878;
-    margin-bottom: 5px;
-  }
-}
