--- src/pages/Contents/MainArea/MainArea.scss
@@ -1,5 +1,0 @@
-.main-contents {
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-}
