--- src/pages/Contents/MainArea/MovieInfo/MovieInfo.scss
@@ -1,17 +1,0 @@
-.movie-info {
-  display: flex;
-  justify-content: center;
-  height: 200px;
-  margin: 0 auto 20px;
-  padding: 0 200px;
-  background-color: #ffff;
-  border-bottom: 1px solid #e3e3e3;
-
-  .movie-info-all {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    margin: 22px 0 23px 25px;
-    text-align: left;
-  }
-}
