--- src/pages/Contents/MainArea/MovieInfo/MoviePoster.scss
@@ -1,8 +1,0 @@
-.movie-poster-all {
-  .movie-poster {
-    width: 165px;
-    margin-top: -54px;
-    border: 2px solid #ffff;
-    border-radius: 5px;
-  }
-}
