--- src/pages/Contents/MainArea/MovieInfo/MovieTitle.scss
@@ -1,13 +1,0 @@
-.movie-title {
-  .title {
-    font-size: 33px;
-    font-weight: 700;
-  }
-
-  .text {
-    width: 780px;
-    margin: 8px 0;
-    font-size: 17px;
-    color: #7f7f7f;
-  }
-}
