--- src/pages/Contents/MainArea/Producers/Process.scss
@@ -1,19 +1,0 @@
-.process {
-  margin: 20px 0;
-
-  .title {
-    margin: 20px 20px;
-  }
-
-  .process-all {
-    margin: 10px 0 10px;
-
-    &::after {
-      content: '';
-      display: block;
-      margin: 0 auto;
-      width: 598px;
-      border-bottom: 1px solid #f0f0f0;
-    }
-  }
-}
