--- src/pages/Contents/MainArea/Producers/Producers.js
@@ -1,112 +1,0 @@
-import React, { Component } from 'react';
-import './Producers.scss';
-
-class Producers extends Component {
-  render() {
-    return (
-      <ul className="producers-all">
-        <li className="producer">
-          <img
-            src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
-            className="producer-image"
-            alt="출연/제작자"
-          />
-          <div className="producer-info">
-            <p className="producer-name main-text">양조위</p>
-            <p className="producer-act sub-text">조연 | 웬우</p>
-          </div>
-        </li>
-        <li className="producer">
-          <img
-            src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
-            className="producer-image"
-            alt="출연/제작자"
-          />
-          <div className="producer-info">
-            <p className="producer-name main-text">양조위</p>
-            <p className="producer-act sub-text">조연 | 웬우</p>
-          </div>
-        </li>
-        <li className="producer">
-          <img
-            src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
-            className="producer-image"
-            alt="출연/제작자"
-          />
-          <div className="producer-info">
-            <p className="producer-name main-text">양조위</p>
-            <p className="producer-act sub-text">조연 | 웬우</p>
-          </div>
-        </li>
-        <li className="producer">
-          <img
-            src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
-            className="producer-image"
-            alt="출연/제작자"
-          />
-          <div className="producer-info">
-            <p className="producer-name main-text">양조위</p>
-            <p className="producer-act sub-text">조연 | 웬우</p>
-          </div>
-        </li>
-        <li className="producer">
-          <img
-            src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
-            className="producer-image"
-            alt="출연/제작자"
-          />
-          <div className="producer-info">
-            <p className="producer-name main-text">양조위</p>
-            <p className="producer-act sub-text">조연 | 웬우</p>
-          </div>
-        </li>
-        <li className="producer">
-          <img
-            src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
-            className="producer-image"
-            alt="출연/제작자"
-          />
-          <div className="producer-info">
-            <p className="producer-name main-text">양조위</p>
-            <p className="producer-act sub-text">조연 | 웬우</p>
-          </div>
-        </li>
-        <li className="producer">
-          <img
-            src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
-            className="producer-image"
-            alt="출연/제작자"
-          />
-          <div className="producer-info">
-            <p className="producer-name main-text">양조위</p>
-            <p className="producer-act sub-text">조연 | 웬우</p>
-          </div>
-        </li>
-        <li className="producer">
-          <img
-            src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
-            className="producer-image"
-            alt="출연/제작자"
-          />
-          <div className="producer-info">
-            <p className="producer-name main-text">양조위</p>
-            <p className="producer-act sub-text">조연 | 웬우</p>
-          </div>
-        </li>
-        <li className="producer">
-          <img
-            src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
-            className="producer-image"
-            alt="출연/제작자"
-          />
-          <div className="producer-info">
-            <p className="producer-name main-text">양조위</p>
-            <p className="producer-act sub-text">조연 | 웬우</p>
-          </div>
-        </li>
-      </ul>
-    );
-  }
-}
-
-export default Producers;
