--- src/pages/Contents/MainArea/Producers/Producers.scss
@@ -1,38 +1,0 @@
-.producers-all {
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-  height: 250px;
-  overflow-x: scroll;
-
-  &::-webkit-scrollbar {
-    display: none;
-  }
-
-  .producer {
-    display: flex;
-    width: 275px;
-    margin: 9px 0 9px 18px;
-
-    .producer-image {
-      width: 55px;
-      margin-right: 10px;
-      border-radius: 5px;
-    }
-
-    .producer-info {
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      margin-left: 3px;
-
-      .producer-name {
-        margin: 3px 0;
-      }
-
-      .producer-act {
-        margin: 3px 0;
-      }
-    }
-  }
-}
