--- src/pages/Contents/MainArea/SimilarThings/SimilarThings.scss
@@ -1,13 +1,0 @@
-.similar-things {
-  margin: 20px 0;
-
-  .title {
-    margin: 20px 20px;
-  }
-
-  .similar-content {
-    display: flex;
-    justify-content: center;
-    margin: 30px 0 10px;
-  }
-}
