+++ src/pages/Contents/Producers/Process.js
@@ -1,0 +1,95 @@
+import React, { Component } from 'react';
+import './Process.scss';
+
+class Process extends Component {
+  render() {
+    return (
+      <section className="process">
+        <h2 className="title medium-title">출연/제작</h2>
+        <div className="process-all">
+          <ul className="producers-all">
+            <li className="producer">
+              <img
+                src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
+                className="producer-image"
+                alt="출연/제작자"
+              />
+              <div className="producer-info">
+                <p className="producer-name main-text">양조위</p>
+                <p className="producer-act sub-text">조연 | 웬우</p>
+              </div>
+            </li>
+            <li className="producer">
+              <img
+                src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
+                className="producer-image"
+                alt="출연/제작자"
+              />
+              <div className="producer-info">
+                <p className="producer-name main-text">양조위</p>
+                <p className="producer-act sub-text">조연 | 웬우</p>
+              </div>
+            </li>
+            <li className="producer">
+              <img
+                src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
+                className="producer-image"
+                alt="출연/제작자"
+              />
+              <div className="producer-info">
+                <p className="producer-name main-text">양조위</p>
+                <p className="producer-act sub-text">조연 | 웬우</p>
+              </div>
+            </li>
+            <li className="producer">
+              <img
+                src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
+                className="producer-image"
+                alt="출연/제작자"
+              />
+              <div className="producer-info">
+                <p className="producer-name main-text">양조위</p>
+                <p className="producer-act sub-text">조연 | 웬우</p>
+              </div>
+            </li>
+            <li className="producer">
+              <img
+                src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
+                className="producer-image"
+                alt="출연/제작자"
+              />
+              <div className="producer-info">
+                <p className="producer-name main-text">양조위</p>
+                <p className="producer-act sub-text">조연 | 웬우</p>
+              </div>
+            </li>
+            <li className="producer">
+              <img
+                src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
+                className="producer-image"
+                alt="출연/제작자"
+              />
+              <div className="producer-info">
+                <p className="producer-name main-text">양조위</p>
+                <p className="producer-act sub-text">조연 | 웬우</p>
+              </div>
+            </li>
+            <li className="producer">
+              <img
+                src="/images/3f2c70fa235fcd74e64cacf38230e85d.jpeg"
+                className="producer-image"
+                alt="출연/제작자"
+              />
+              <div className="producer-info">
+                <p className="producer-name main-text">양조위</p>
+                <p className="producer-act sub-text">조연 | 웬우</p>
+              </div>
+            </li>
+          </ul>{' '}
+        </div>
+      </section>
+    );
+  }
+}
+
+export default Process;
