+++ src/pages/Contents/SimilarThings/SimilarThings.scss
@@ -1,0 +1,44 @@
+.similar-things {
+  margin: 20px 0;
+
+  .title {
+    margin: 20px 20px;
+  }
+
+  .similar-content {
+    display: flex;
+    justify-content: center;
+    margin: 30px 0 10px;
+
+    .similarcontent-all {
+      display: flex;
+      justify-content: space-between;
+      flex-wrap: wrap;
+      width: 598px;
+
+      .similarcontent {
+        display: flex;
+        flex-direction: column;
+        padding: 0 5px;
+        margin-bottom: 20px;
+
+        .similarcontent-poster {
+          width: 139px;
+          border-radius: 4px;
+        }
+
+        .similar-title {
+          margin: 7px 0 5px;
+        }
+
+        .similar-star {
+          margin: 2px 0;
+        }
+
+        .similar-genre {
+          margin: 3px 0;
+        }
+      }
+    }
+  }
+}
